Added a NonRenderedEntity, subclass of StaticEntity, which not only does not update, but does not render.  It is necessarily interactable. Both Door and Asteroid extend NonRenderedEntity

+++ src/com/arnopaja/supermac/helpers/SuperParser.java
@@ -196,7 +196,7 @@ public abstract class SuperParser<T> {
         return json.getAsJsonPrimitive(name).getAsInt();
     }
 
-    protected static void addInt(JsonObject json, String name, float i) {
+    protected static void addInt(JsonObject json, String name, int i) {
         json.addProperty(name, i);
     }
